LawPy/requirements.txt
django==4.2.9
pymongo==4.6.2
python-dotenv==1.1.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
mongoengine==0.27.0 